+++ src/providers/WestWorkspaceDataProvider.ts
@@ -3,9 +3,9 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { WestWorkspace } from '../WestWorkspace';
 
-export class WestWorkspaceDataProvider implements vscode.TreeDataProvider<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem> {
-	private _onDidChangeTreeData: vscode.EventEmitter<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | undefined | void> = new vscode.EventEmitter<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | undefined | void>();
-	readonly onDidChangeTreeData: vscode.Event<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | undefined | void> = this._onDidChangeTreeData.event;
+export class WestWorkspaceDataProvider implements vscode.TreeDataProvider<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | WestWorkspaceEnvTreeItem | WestWorkspaceEnvValueTreeItem> {
+	private _onDidChangeTreeData: vscode.EventEmitter<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | WestWorkspaceEnvTreeItem | WestWorkspaceEnvValueTreeItem | undefined | void> = new vscode.EventEmitter<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | WestWorkspaceEnvTreeItem | WestWorkspaceEnvValueTreeItem | undefined | void>();
+	readonly onDidChangeTreeData: vscode.Event<WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | WestWorkspaceEnvTreeItem | WestWorkspaceEnvValueTreeItem | undefined | void> = this._onDidChangeTreeData.event;
 
   constructor() {
 
@@ -15,14 +15,14 @@ export class WestWorkspaceDataProvider implements vscode.TreeDataProvider<WestWo
 		return element;
   }
 
-  getChildren(element?: WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | undefined): vscode.ProviderResult<WestWorkspaceTreeItem[] | WestWorkspaceMiscTreeItem[]> {
+  getChildren(element?: WestWorkspaceTreeItem | WestWorkspaceMiscTreeItem | WestWorkspaceEnvTreeItem | WestWorkspaceEnvValueTreeItem | undefined): vscode.ProviderResult<WestWorkspaceTreeItem[] | WestWorkspaceMiscTreeItem[] | WestWorkspaceEnvTreeItem[] | WestWorkspaceEnvValueTreeItem[]> {
     if(element === undefined) {
       if(vscode.workspace.workspaceFolders) {
         const items: WestWorkspaceTreeItem[] = [];
         for(let workspaceFolder of vscode.workspace.workspaceFolders) {
           if(WestWorkspace.isWestWorkspaceFolder(workspaceFolder)) {
             const westWorkspace = new WestWorkspace(workspaceFolder.name, workspaceFolder.uri);
-            const item = new WestWorkspaceTreeItem(westWorkspace, vscode.TreeItemCollapsibleState.None);
+            const item = new WestWorkspaceTreeItem(westWorkspace, vscode.TreeItemCollapsibleState.Collapsed);
             items.push(item);
           }
         }
@@ -34,9 +34,30 @@ export class WestWorkspaceDataProvider implements vscode.TreeDataProvider<WestWo
       return Promise.resolve([]);
     } 
     
-    if (element instanceof WestWorkspaceTreeItem) {
-
+    if(element instanceof WestWorkspaceTreeItem) {
+      // Get Zephyr environment variables
+      const items: WestWorkspaceEnvTreeItem[] = [];
+      for(let key of WestWorkspace.envVarKeys) {
+        const envItem = new WestWorkspaceEnvTreeItem(element.westWorkspace, key);
+        items.push(envItem);
+      }
+      return Promise.resolve(items);
     } 
+
+    if(element instanceof WestWorkspaceEnvTreeItem) {
+      // Get Zephyr environment variables
+      const items: WestWorkspaceEnvValueTreeItem[] = [];
+      let values = element.westWorkspace.envVars[element.envKey];
+      if(values) {
+        for(let value of values) {
+          const envValueItem = new WestWorkspaceEnvValueTreeItem(element.westWorkspace, element.envKey, value);
+          items.push(envValueItem);
+        }
+      }
+      
+      return Promise.resolve(items);
+    } 
+
     return Promise.resolve([]);
   }
 
@@ -79,4 +100,28 @@ export class WestWorkspaceMiscTreeItem extends vscode.TreeItem {
 	}
 
 	contextValue = 'west-workspace-misc';
+}
+
+export class WestWorkspaceEnvTreeItem extends vscode.TreeItem {
+  constructor(
+		public readonly westWorkspace: WestWorkspace,
+    public readonly envKey: string
+	) {
+    super(envKey, vscode.TreeItemCollapsibleState.Collapsed);
+    this.description = westWorkspace.envVars[envKey].length === 0 ?'[not set]':'';
+    this.tooltip = envKey;
+	}
+  iconPath = new vscode.ThemeIcon('variable');
+  contextValue = 'west-workspace-env';
+}
+
+export class WestWorkspaceEnvValueTreeItem extends vscode.TreeItem {
+  constructor(
+		public readonly westWorkspace: WestWorkspace,
+    public readonly envKey: string,
+    public readonly envValue: string
+	) {
+    super(envValue, vscode.TreeItemCollapsibleState.None);
+	}
+  contextValue = 'west-workspace-env-value';
 }